+ Started a Collaboritve Game Project
I'm very excided to announce that for my 2019 Christmas Vacation project,
I am collaborating with the lovely and talented [@pointyearsillustration]
(https://www.instagram.com/pointyearsillustration/) on a short, old-school
point-and-click adventure game.  We're still hashing out the details of the
story, but the game will follow a boy preparing for an adeventure who has
to overcome a series of ridiculous obstacles and setbacks before he can be
on his way.  We're combining simple point-and-click game design with Sara's
signature style aiming for a short, lighthearted adventure that's "done"
before Jan 6th.

For this project I'm using the GoDot Engine, which I chose due to it's excellent
documentation, ability to develop cross-platform (I'll be using Linux and OS X),
and its easy exporting to the web.  I've never _really_ used GoDot before (except
for that [one thing](https://github.com/Dorthu/Tanks-of-Freedom/tree/feature/objectives)
I did years ago), but the docuemntation should be enough to get me going quickly,
and exporting to the web is crucial to making this something people might actually
try.  If we have time, I'm going to look at making this mobile-web friendly as well, but
for the initial goal I just want a finished game that works in brower on desktop.