+ Stared a Dialog System
Most of the technical elements of this game are relaitvely simple, and
moreover require graphics to work on.  Instead of making test graphics,
I'm focussing for now on one of the more code-heavy pieces; the dialog
box system.  While this can likely be achieved with a simple GoDot text
field, I'd really like to get those cool effects you see in other games,
where letters appear one at a time, and some words appear or animate
differently.  This entire game is going to be light on animation to save
time, so having some in the text might compensate.

So far I've added a DialogController that can create Dialog Boxes, and
can keep them in a static position on the screen regardless of where the
camera is panned to.  That doesn't sound (or look) like a lot, but I'm
also still learning this engine, and it took a decent amount of reading
docs to figure out how to do it right.