+ Working Dialog Box
I got just the basics done; so far you can set text for the dialog box
and have it appear slowly.  The input text isn't parsed into different
pieces yet, but it will be.  As far as an initial pass goes, though,
this works.  The next step is parsing, which will split the input string
into different pieces that should be disaplyed in different ways (for
instance, a string might have some words that appear slowly, then some
that appear all at once).