+ Working Text Animations
After a bit of work on the dialog interface, I now have three different
text "styles" that can be used when displaying dialogs.  They're
governed by the `|` character in the text string, and the character
following it.  For example, I string like `This is normal.` would appear
at the normal text speed, but a string like `|_This is slow.` would show
up much more slowly.  You may've seen this kind of effect in other games
(I'm lifiting it right out of Undertale, personally, but take your
pick), and I find it conveys a lot of emotion with little/no animation
(which is what this game is going to have).  For example, a string like
`I need you to |!stop |!it |!now.` would appear normally until "stop",
then "stop", "it", and "now" would all appear at once, which conveys the
emphasis put on the words when spoken.  I've found this to be equally
impactful and intuitive; nothing in a game told me that the way the text
appears conveys this feeling, but I got it immediately anyway.

My next step is to make objects that can live in scenes and that can
display dialogs when clicked on; I intend for the editor to have a slot
for putting their descriptions in, and that said descriptions are parsed
using this formatting.  This should make developing rooms pretty quick.

In other news, Sara and I have come up with the first two objectives -
to fetch a backpack by convincing the plant that is the hook it's hung
up on to unfurl and allow its taking, and to appease some kind of
monster so that you can take a potion of some sort (which will itself
require attracting a bumblebee to get the item the first monster
wanted).  I guess I need to put some thought into an inventory pretty
soon too.