+ Music, concepts, & progress
Yesterday, I reached out to the extremely talented [Josh
Sager](https://www.youtube.com/channel/UCPcdjB0r0gU13m0oSizEdSA) (a
friend who I've worked with for some time on unrelated things) asking
if he'd have any interest in making some ambient music for this project
(he's been doing a lot of ambient music lately with some really
facinating technology) - and today, he text'd me out of the blue to
share the track he put together.  I haven't gotten to implementing it
yet, but it's included here in the project now, and should be in soon.

I also wanted to dive into the concept work Sara and I have been doing;
chiefly this:

![concept](https://raw.githubusercontent.com/Dorthu/2019-christmas-adventure/master/meta/concept_1.jpg)

This is the result of Sara and my original disucssion of this project,
including going over the core mechanics, the structure of the story, and
what this game should aim to be.

![concept](https://raw.githubusercontent.com/Dorthu/2019-christmas-adventure/master/meta/concept_2.jpg)

This is the concept for the first quest, wherein the hero must retrieve
their backpack from a living hook that's got a hold on it.

I thought these might be interesting to share to break up an otherwise
dull dev blog, where I might've just written "dialogs can now appear
when you click, and can be dismissed."  I next need to show the dialog
for only clicks _on_ objects.