+ Added Room Transitions
Now we're getting somewhere - when you mouse over a door, an arrow
apperas indicating that you can go through, and when clicked, you do!
Right now the second room is just a reversed version of the first room,
but that's fine; it's enough to prove that it's working.

Up next, I'd like to make the mouse tracking more simple (i.e. using the
global mouse position instead of trying to record it) and maybe add some
additional features to dialogs (like multiple lines, text completion on
click, etc).