+ Going Away Tomorrow
Tomorrow, we're flying down to Florida to visit with family.  I'm really
looking forward to the trip, and I hope to continue working on this
project in the margins while I'm there; there won't be much time left
after I get back to finish is.

This update includes improvements to mouse tracking (using global mouse
position instead of managing it ourselves) and improvements to the
dialog system (namely, clicking completes the current dialog box, and
multiple lines are allowed to be displayed now).

Hoping for a safe, uneventful flight; next time I post here the weather
will be a lot nicer.