+ Safe Travels
We made it to Florida yesterday, but it was a long day of traveling and
visiting, and I went to bed early - no time to work on this at all.
That said, the weather down here is beautiful, and before I head back
down to the pool, here's an update that adds arbitrary-length dialog
boxes, broken into three-line segments.

I had some issues with GoDot's limitations around export types - namely,
I can't define `export dialog_lines: DialogTextPool` as `DialogTextPool`
is a custom type.  I'd expected to be able to define exported variables
in my custom class, but that's not a thing.  This is going to complicate
conditional dialogs; adding a global game state is high on my list, and
I was figuring on exporting a list of "(game state, dialog text)"
objects for `Clickable`, but since I can't export custom types I might
have to be more creative.

Anyway, my son's getting impatient, so I'll think about it later.