+ A Wonderful Day Out
Sara and I got the chance to go out by ourselves overnight on one of the
islands, and it was a great time.  We got to walk the beach as the sun
set over the gulf, have some tropical drinks, and spend some quality
time together without interruption.  It was awesome.

After getting back, I put in logic to handle "openable" objects; things
like cabinets that can be opened to reveal their contents.  I think it's
finally time I tackled the inventory, as that is needed for most of the
rest of what I want to do.