+ Coffee Shop Inventory Moment
Having taken some more time to ourselves, we left the house in the
evening to go work at a coffee shop, where I got the first half of the
inventory system working (the part where you can collect stuff) while
someone played their "chili song" on the acoustic guitar in the
background.  Clever diddy about how chili for supper is the true test of
love.  I've been to a few open mic nights, and this is amongst the
better ones.

To keep the interface simple, Sara and I decided that items in the
inventory won't have a specific description; instead, you can
(optionally) get a dialog when you collect the item, but that's it -
subsequent clicks on the item will use it.  I think it's important that
this game is accessible to everyone, and I don't want to bog it down in
mechanics that don't really add anything.

Someone else is on the mic now, and he's really good.