+ Potion Quest
Sara and I spent some time tonight finishing the quests and concept
drawings for this project so she can finish the assets.  Here's the
second quest:

![Second
Quest](https://raw.githubusercontent.com/Dorthu/2019-christmas-adventure/master/meta/concept_3.jpg)

This quest happens alongside the remaining two - they can be taken in
any order, and even simultaneously.  They're all to retrieve some item
that you need in order to leave for your adventure - one of those items
is a potion.

The potion is found in the cellar - you need to pilfer the key before
you can even get in.  Once there, a "grump" (some kind of creature) is
drinking the potions, and you have to give him something else to drink
to get the potion you want.  That "something else" is some of Dad's
liquor, which is too high up on a shelf for you (a child) to retrieve on
your own - so first you need a stool.  Once the liquor is traded for the
potion, this quest is complete.

This is the only quest that goes into the cellar, and the only time you
need to interact with the grump, but hold on to that stool - it'll come
in handy in another quest later on.

In this update, I added the _real_ rooms - we finished the entire
outline of the game, so I now know where everything is and how it
connects, and all I need is for it to all be created.  Two of the four
rooms still need a sketched out background, so they're blank (but the
doors exist at approximately the right place and work).  I likely won't
be using these rooms for continued development, as I still need my test
assets in order to finish the global game state.