+ Locked Doors
Tomorrow is our last full day in Florida, and I already miss the
weather.  I'm sitting outside on the porch working on the game right
now, as I have with many a vacation project before this one, and it's a
very comfortable 71, which is awesome in December.

To follow up on the potion quest, I added in the `LockedDoor` class;
this class is maybe more generic than it needs to be seeing as there's
only one in the whole game, but what can I say?  It is also the first
piece of the "game state" tracking, currently being done with a
Dictionary at `MastterController.game_state` that holds arbitrary
values.  I assume future classes that care about state will use this as
well.