+ Items in Cabinets
Part of this game is going to be searching through the house for items,
and as such some items will be contained within cabinets.  I've added
code to `Openable` to handle this; now, one item can be contained within
an `Openable`, and will be revealed when it's opened.

Unfortunately, Godot doesn't handle propagation of click events in
overlapping `Area2D` instances like it does other events; it looks like
unless I write some code to handle that myself, I'll need to make sure
that `Openables` that have items in them don't have a description of
their own.  I think that's probably okay.