+ Lofty Goals
When I originally pitched this project to Sara before the vacation
started, I'd figured this kind of game would be a breeze to make, even
in an unfamiliar engine and toolset.  That hubris is comin' round to
haunt my now, as it's clear this won't be even close to finished by the
end of Sunday.  So.. here's what I'm thinking.

Sara is excited for this project, and committed to finishing it.  I am
too; considering her work is so popular, this is a game people might
actually play!  So I'm going to post what I have done before the
vacation is over, and then we're going to keep it going and see what we
can do,  I think this has the potential to be a really awesome, if
short, adventure game.

It's not the first vacation project that I didn't have time to finish,
and it won't be the last.