+ Last Day in Paradise
We're leaving Florida today, and it's raining.  That's given me some
time to get things together for the backpack quest, which I'd like to
have "working" before I post the game.  I made quick, temporary graphics
for all the required items, and just wrote the code for the
`LockedOpenable`, which is like an `Openable` but requires an item to
open.  This is what the branch-hook that's holding your backpack is
implemented as; it doesn't really "open" like a cabinet, but the branch
unfurling into the water and revealing the backpack works the same way -
and it looks pretty good, if I do say so myself.

Sara's going to work on sketches for the remaining two rooms, then she's
going to start trying to replace these test graphics.  I still need to
figure out how to submit things to itch.io, which is where I expect this
to end up when it's "done."