+ The Longest Night
Last night we got to the airport to catch our flight home at 5:30 to
find it delayed from 7:40 to 10:05.  While waiting, it was further
delayed, _twice_, before finally being cancelled at 11.  The next flight
we could book even close to home was on Monday, so I had to take another
day off, Pat's missing another day of school, and we've got another
flight to get through.  Also, I cannot recommend flying with Frontier -
I get that things happen and flights get cancelled, but the way they
handled it was terrible; they didn't even announce the delays at the
terminal, and the cancellation came up on my phone long before they said
anything.  Horrible.

But it does give us more time to finish this project up!  Or, at least,
"finish" it up.  Sara has the last two rooms sketched out, and I got the
backpack quest fully implemented.  It even took a customer `Clickable`
class for the `Sink`, which exchanges an item for another item.  I might
make that more generic, if I decide I need it elsewhere.

At this point, I need an opening dialog to run, the last graphic for the
kitchen imported, to setup the initial game state, and to figure out how
to export/host the final product.

So, we've got a day.  I think we can do it.