+ Itch.io
This project is going to be hosted on [itch.io](https://itch.io), and I
did the initial export and configuration this morning.  Turns out, you
need a title for a game when setting it up on there... and we didn't
think of that.

Anyway, the initial draft is up, and it kinda works?  It's not public
yet, and with good reason; the viewport is way too long on the exported
game, and at the same time the game itself is not scaled (so it's
actually rendering at a height of 600px).  I think it's an issue in the
export settings, so I'm gonna fiddle with it.  To be honest, it exported
and runs on a webpage on itch.io, and that's more than I expected from
the first draft.

Anyway, our plane home leaves at 2:30, and with any luck it'll be an
uneventful flight.  Sara is going to use it as an opportunity to
finish up the remaining graphics, I'm going fiddle with the export
settings, and we'll hopefully have this uploaded and public before too
long.