+ Friday Jamz
In an effort to finish up this project, Sara and I have decided to make
Friday night "Game Jam Night."  Tonight, we got a pair of `bang` energy
drinks and some long-winded YouTube videos, and we're hanging out
getting this done!  Weekend warriors!

I got a custom export tempalte configured so that I don't need to
manually mangle the exported HTML to get it to render how I want on
itch.io, which is nice.  Sara's working on finishing up the kitchen
background.  We've only got a few more resources left before it's
"releasable" as an in-progress demo game.  Once that's done we can focus
on the remaining quests, which should be pretty light on resources and
heavier on the coding effort; luckily I can get those going while Sara
does the color on the existing background, and hopefully we can converge
on a finished game together.

I'm glad this is still happening.