+ Almost Ready
Last night's Friday Night Jam was very successful, with the kitchen
being finished, many of the remaining resources being completed, and the
a huge amount of cleanup work being done throughout the project.  I got
rid of almost all of the temporary resources, redid the folder
structure, and excluded the metadata folder from the game (which should
reduce the file size significantly).

This is all we need for an "initial release":

 * An arrow for transitioning rooms
 * A background for the inventory

Going forward, we'll also need to finish up all the room background and
get the last three quests finished, add some sound effects, maybe even
spice up the intro.. but for getting something public on itch, that's
all that's left.