+ No More Placeholder Graphics!
Sara's finished at least a sketch of all the graphics for the game,
meaning that I've been able to remove all the placeholder graphics I
made from the project - we're now running entirely with the real deal!

We're going to sleep on it, but right now I'm leaning towards uploading
this to itch and making it public tomorrow.  Very exciting!