+ Launched!
Maybe more of an "early access" launch, really, but we're live on
itch.io at https://dorthu.itch.io/working-title and I'm pumped!  Also
naming things is hard!

This is obviously a huge milestone for the project, but we're not even
close to done yet - we've still got three more quest lines to put into
the game, all of the backgrounds need to be finished, and I have
aspirations for "cutscenes" in the intro and outro of the game.  Lots to
do future friday nigths, that's for sure.

I'm still excited to see if anyone tried it, and if they do, what they
think.  While I've been making games off and on for years, I've seldom
gotten something so far along that it could be "shipped" somewhere, and
certainly never a point-and-click game.