+ Dialog Profiles
This Friday Night Jam, Sara had prepared a few profile pictures for use
in dialog boxes, so I went ahead and implemented them while drinking
White Claw because I'm not too old to be hip yet!  Also they're good.

This update, and others going forward, should be published to itch in
the next day or so; we're gonna keep going until we're asleep, so it
won't ship tonight.

To control the dialog profiles, I extended the parsing to include `|P`
as a token that controls the profile for the line - only the profile on
the first line of a dialog is respected.  That's not perfect, but it fit
nicely into the existing system, which I don't expect to need to further
extend.  The token following `|P` and before the next `|` is the name of
the profile image in `res://img/dialog`, meaning that as new profile
pictures are added I can simply change the text prompts to include them.

I went through most of the text in the game and added profiles where
appropriate.