+ Godot 3.2
Yesterday, [Godot 3.2 was
released](https://godotengine.org/article/here-comes-godot-3-2).  That's
super exciting, as it has a slew of new and improved features that have
made working on this project a lot easier; for instance, the node editor
was totally overhauled so that linking functions to objects isn't a pain
in the neck anymore.  I upgraded immediately, and it works great.

I got the bumblebee working, finally.  Turns out the angle function I
was using was returning the angle from the X axis, in a range of -90-90
degrees (but in radians of course), and the real solution was to think
of everything as Vectors instead of Points and Angles.  Now, the bee has
a Vector2 of velocity, and uses the desired direction toward its goal as
a modifier on its velocity vector.  It works pretty well.  Some tweaking
to add acceleration when fleeing from the mouse and lose it when
turning, and it's a fairly natural motion.

I also started on the stool, another big part of the game; there are
several things planned that the player can't reach without it.  I needed
a way to show the player where the stool could be placed without having
to have them click around randomly, but also without being too obvious
and intrusive - I settled on a phantom-stool that appears (and shimmers
a bit!) when the stool is in the inventory.  That's all coded up (I
wrote my first Godot shader, even!) and will be included in the next
release, so check it out.