+ Bumble Quest
This Friday Game Jam, I wasn't really feeling it.  Maybe it was a
particularly busy week at work, maybe it was rewiring some lamps earlier
this evening, or maybe I was just tied, but I sat down with my Bang!
energy drink and didn't have any ideas.

So, while the last episode of Netflix's pretty..bad.. Dracula series
played in the background, I started on finishing the bumble quest.  I
made pickup items that depended on game state, and once I was done I
noticed the warning that a subclass was overriding a parent class'
variable - lo and behold I'd done the same thing last week, and so
similarly that some of the variables were named the same!  At least I'm
consistent.

That aside, it was a hugely successful night!  Sara finished the bumble
graphic, and I finished the coding - now the game isn't over until you
have a backpack _and_ honeycomb.  Here's the concept:

![Bumble
Quest](https://raw.githubusercontent.com/Dorthu/2019-christmas-adventure/master/meta/concept_4.jpg)

This is a fairly simple quest, but with a one-off mechanic; when you
open the kitchen cabinet, a bumblebee flys out with the honeycomb you
need.  It avoids your mouse, and continues buzzing around the room until
you offer it a flower.  You make this offer by selecting the flower from
your inventory, but not actually clicking anything (causing the flower
to follow your mouse).  Once the bumble reaches the flower, it'll trade
you for the honeycomb it's carrying and fly off.

I realized tonight that the mouse interactions were still wrong
sometimes, and by drawing what I thought was the mouse position, I
realized why: the mouse position wasn't being moved to account for room
panning!  Once that was sorted, it behaves exactly as I'd expect.