+ Imp Quest
This Friday Night Jam, we've got the imp quest "finished!"  And by
"finished" I mean that  most of the graphics are done, and all of the
dialog is temporary - but it works!  This is a great time to introduce
Frank:

![Frank](https://raw.githubusercontent.com/Dorthu/2019-christmas-adventure/master/img/cellar/imp.png)

I don't know if Frank is going to be his final name, but this little
imp guy is central to the second quest - he guards the potion you need
to leave the house.  Frank likes liquor, so you've got to give him the
bottle in exchange for the potion.

Frank is eventually going to have a rotating dialog pool, where some of
it includes hints on what he wants, and some of it builds the characters
of the game.  These might even get put in before we publish the next
update.  But for now it's tomorrow, and Brett's heading out after this
YouTube video is done.