+ The Final Quest
The original concept for this game included the "tutorial quest" of
getting your backpack, and then three more things to collect, each a
"quest" of their own, which you could do concurrently.  Tonight, I
started on the last of them.

![the final quest](https://raw.githubusercontent.com/Dorthu/2019-christmas-adventure/master/meta/concept_5.jpg)

This quest involves getting a sword and shield from over the mantle of
the fireplace.  In order to reach, you need the stool set before the
first, but there's a problem; a domovoy lives in the fireplace, and he's
not having it!  In order to get the stool set in place, you need to
appease him (with cookies!).

This seemed pretty simple to implement, until I did it and realized it
would be a lot more impactful to actually have the stool (with the cool
shadow effect) in place while the domovoy objected; that required
chaining dialogs, which didn't work as easily as I'd've hoped as
functions aren't first-class entities in GDScript (at least not until
4.0).  As such, some additional infrastructure was needed in the
MasterController to support dialog_finished callbacks.

We're very close to a fully functional game now; at this point, I mostly
need graphics, and to get some cutscenes in place.  I'm also looking at
recording and implementing sound effects sometime soon.