+ All Quests Complete
This is a huge milestone for the game: *ALL QUESTS ARE FINISHED*!

That's right, as of this commit, the entire arc of the game is playable!
That's not to say the game is _finished_; there's still plenty to do
before it's done, but it's done-ish!

This time, I added the quest for the Sword & Sheild (as described in the
last post), and tweaked a few existing things.  Notably, I added a cool
particle effect to Frank in the cellar, which I think goes a long way to
bringing that room to life.  I'd like to add at least one effect to
every room to add some extra interest to the scenes; the bumblebee in
the kitchen is unexpected and brings the room to life, the fire in the
living room adds some motion and interesting shadows (more to come when
the room is finalized, as I intend to have flickering shadows in the
stones lining the hearth as well), and now the cellar has a particle
effect for Frank sleeping.  The only room without something is the
bedroom.. which I'll have to think about.

The next big piece for me to work on is sound effects; I think that
having a small sound clip for most interactions (or at least most
interactions that amount to something) will really add some polish.  We
also intend to add cutscenes, one in the start and one at the end,
which should serve to frame the story nicely and give you some
additional impetus/resolution.

I don't want to set unreasonable expectations, but I'm going to push to
have this game finished by the end of March.  Just a _little_ later than
the initial goal of "before January 6th".