+ Accessiability
Last weekend, after publishing the update, I was feeling pretty good.
This game was really coming together.  I was excited to see someone
actually play it.

Who better than _my dad_.

Now, I mean no offense when I say that my dad is not the best or most
experienced computer user on the planet.  In fact, he may well've never
played a proper video game before, ever.  So when I asked him to give
our game a try, it was treading new ground for him.  And that was
perfect.

Watching him get through the first few tasks was very eye-opening for
me.  I thought I was designing this to be accessiable to people who
aren't familiar with commonalities of video games, but it turns out I'd
made some assumptions I didn't notice, like that you should click on
things, or moving the mouse is part of gameplay.

To that end, this week's update is mostly focused on making the systems
of the game less mysterious to players that maybe aren't familiar with
video games in general.  There's now a "Click to Continue" prompt that
appears on dialog boxes, but only a second or so after all the text has
appeared (which means players that know what to do will probably never
see it).  We added a small tutorial section that instructs a new player
how to look around a room (by moving the mouse) and how to interact (by
prompting a click).  This has a secondary benefit of guiding players to
the right to begin with, which will lead them right to the backpack (and
the first puzzle) _and_ almost gaurentees their mouse will go past the
front door, displaying its exit arrow and hopefully prompting a click
(to show players where their ultimate goal is, and how to get a reminder
of what needs to be done to get there).

I think the most eye-opening thing my dad revealed during his
playthrough, though, was a total design oversight on my part: when asked
to find water for the branch, he found his way into the kitchen, then
saw a pitcher sitting on the counter and clicked it.  He was then
confused as to why nothing happened.  I hadn't even considered this, but
I'd accidentially introduced one of the worst possible pitfalls in
puzzle desing - a solution that makes sense, but doesn't work for no
reason!  This one was an easy fix (now there's a prompt telling you the
pitcher is too heavy, we might remove it entirely), but it's also got me
looking over the whole game to make sure there are no more logical
solutions that we put in the game that just don't work.

The last thing I noticed was that it was hard for him to spot the cup.
I moved it down the counter, but I also added a mouse-over hint icon
that highlights all clickable elements on the screen, which should make
it very obvious what you can interact with to avoid that situation.

On the positive side, with the few mechanics the game has described to
him, my dad actually made it through the first quest.  It was very
exciting to see him think through the puzzles (he gave the branch the
empty cup, triggering the "the cup is empty" message and reasoned that
the sink was the right place to go!), and it gave me some real
confidence that this is going in a reaonsable direction.

Once we get a little more done, I intend to seek out more testers of
various levels of video game experience.  Hopefully we can get the whole
game very welcoming to anyone who would play it; the intended audience
isn't limited to just gamers, after all.