+ The Final Stretch
I'm super excited to say that with tonight's update, Sara and I both got
something _huge_ finished!  We're on the last stretch of this project
before we can call it "done!"

Sara finished the living room - a massive effort on her part - and it's
been imported into the engine.  It looks _awesome_.  That's also the
last room done, and the remaining graphics required for this are very
small.

An interesting fact here is that in addition to the entirety of the
background, Sara also did her first normal maps (for the fire in the
fireplace).  They look really good, and the dynamic fire and shadows
really bring the room to life!

Less immediately noticiable is what I did - I've finished mobile
support!  Check the game out on itch on your phone - it works pretty
much completely!

Mobile support was a pain, mostly because the feedback loop had a lot of
steps - export the game to a folder I'm serving on my LAN, load that
page on my phone, wait for the game to load, see what worked.  Testing
it on the PC wasn't really cutting it, as it didn't represent the
reality of a phone screen (and there was no touch support).  Some of the
functions I was depending on were pretty poorly documented, too -
nowhere in the docs does is say that `warp_mouse` can't move the mouse
in HTML5 games, for instance (although I understand why it can't).

So, what's left?

Well.. not a lot, to be honest.  We've got to proof all the dialog in
the game, we've got to discuss and add cutscenes for the beginning and
the ending, and most importantly (and most difficultly) we've got to
name it.  There are a few odds and ends besides, but nothing major.

I feel really good about this project.  I can't wait for it to be
complete.