+ Cutscenes
Time for another concept art post!

![cutscenes](https://raw.githubusercontent.com/Dorthu/2019-christmas-adventure/master/meta/concept_6.jpg)

Tonight, we started on the cutscenes.  There'll be two; the first will
be at the very beginning, a dream of what happened right before the game
starts to frame the story.  The second happens when you get all the
items and finish the game, and it's already implemented!

The goal here was to take the existing objective framing and bolster it
with some plot.  The opening dialog says "Dad left last night and isn't
back yet" - who's Dad?  Where'd he go?  Why are we worried?  That's what
the dream will cover, at least partially.

When you finish the quests and leave, the exit cutscene is just you
leaving - this game leads into the story of the book Sara's writing, so
the ending cutscene doesn't cover much except giving more resolution
than a dialog box.

Implementing these wasn't actually that bad; I was worried that my
master controller wouldn't like sharing with a cutscene orchestrator,
but with a simple `cutscene_mode` I got it playing nice.

Once these are done, plus a few odds and ends and some mobile polish,
this game is finished.  Very exciting!