+ Sometimes, I'm Wrong
Last week we did a first draft of the ending cutscene using an idea I
had to animate it simply.  If you haven't seen it, go beat the game real
quick and give it a look - it's.. bad.  And not because of the art; the
aesthetic just doesn't match.

I should've listened to Sara when she told me a few still images would
do the job better.  We're moving in that direction now.

This weekend's update is all about mobile controls.  When I first got
this running on mobile, I thought the "edge of screen mouse pan" worked,
but it was immediately obvious when anyone else tried it that it was
unintuitive - you just want to click-and-drag to pan on a touchscreen.
So that's what we're doing now, and it's effectively an entirely
different interface.

Implementing click-and-drag panning wasn't super hard (fortunately my
initial camera code was pretty easy to modularize); the harder part was
making the rest of the game work _with_ that input style.  For example,
if you happened to tap on something you could interact with, you
immediately got a dialog, and you couldn't pan - super disruptive when
you accidentially tapped something when attempting to just look around.
The fix for this ended up being that on mobile, things are interacted
with on mouse _up_ instead of mouse _down_, and we track if you were
panning the screen before the mouse up to know if it should count or
not.  Fortunrately again, the interaction code was simple enough to
allow a function call to replace the existing condition without much
fuss (although more places than I remembered repeated that check).

With that sorted, interacting with items became weird; it was more
natural to click-and-drag inventory items, especially when everything
else worked on mouse up now.  So again, mobile's input scheme diverges -
on mobile, you drag inventory items onto interactable entities instead
of selecting them and then selecting the entity.  I think this feels
prety natural.

So, I was also wrong about mobile input, initially.  But being wrong is
okay, because it's how we learn.  I've been playing The Witenss lately,
and it makes wrongness a mechanic in some ways - it presents you with a
simple puzzle to teach you, but as the puzzles get harder, your
assumptions get proven wrong, and in doing so you get closer to
understanding.  I was wrong about cutscenes, and I was wrong about
mobile input, but in being wrong I learned how to do it better.

Did we waste a night of work?  Maybe.  Or maybe we spent that night
learning how to make the game even better.